chore: update action.yaml for deno compatibility, improves usability
