--- registry-login/src/cleanup.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-/bin/deno run /app/post.ts
