--- registry-login/src/deno.json
@@ -1,5 +1,0 @@
-{
-  "tasks": {
-    "dev": "deno run --watch main.ts"
-  }
-}
