--- registry-login/src/deno.lock
@@ -1,16 +1,0 @@
-{
-  "version": "3",
-  "redirects": {
-    "https://esm.sh/@actions/core": "https://esm.sh/@actions/core@1.10.1"
-  },
-  "remote": {
-    "https://esm.sh/@actions/core@1.10.1": "4c82ced837ff42a79580a6e3d6869816a87f3d9689376dc3f05f44d11bb7e9d8",
-    "https://esm.sh/v135/@actions/core@1.10.1/denonext/core.mjs": "9558f5f1d14a8307945016c1f9df4040c6e9bea64e10cac331fe0bb241f54937",
-    "https://esm.sh/v135/@actions/http-client@2.2.0/denonext/http-client.mjs": "4aca5cb3d771ad4d6c47fae667a9a659aeee953527da3d7ae733fcf7beb82cb0",
-    "https://esm.sh/v135/@actions/http-client@2.2.0/denonext/lib/auth.js": "bdf8aa70bd820a1a4fb3f9f71728e72b1d5aca36406598354b89e6a944d0aec5",
-    "https://esm.sh/v135/@fastify/busboy@2.0.0/denonext/busboy.mjs": "96f197a55accd534085d32a75861c278803d6c6bb68dcf7925fea7f3139280e6",
-    "https://esm.sh/v135/tunnel@0.0.6/denonext/tunnel.mjs": "0a5bc75c6c9ebb994955935269d2462da6de3f5238df9e21e396ba21599ec6dc",
-    "https://esm.sh/v135/undici@5.27.2/denonext/undici.mjs": "30fd6e093121c9059bc789b1ff8e48a5f70fcdd9a759d36a84d1b21555f8c1cb",
-    "https://esm.sh/v135/uuid@8.3.2/denonext/uuid.mjs": "c0da38266b24ac79d62b02290f17caba7e75d078f6771bbe8fb61bdef60f837c"
-  }
-}
