--- registry-login/src/main.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-/bin/deno run /app/main.ts
