--- registry-login/src/main_test.ts
@@ -1,6 +1,0 @@
-import { assertEquals } from "https://deno.land/std@0.208.0/assert/mod.ts";
-import { add } from "./main.ts";
-
-Deno.test(function addTest() {
-  assertEquals(add(2, 3), 5);
-});
